feat(frontend): add prefix tag to tokens in ManageModal
