+++ src/frontend/src/icp-eth/components/tokens/ManageTokens.svelte
@@ -36,6 +36,8 @@
 	import { exchanges } from '$lib/derived/exchange.derived';
 	import { tokensToPin } from '$lib/derived/tokens.derived';
 	import type { ExchangesData } from '$lib/types/exchange';
+	import Tag from '$lib/components/ui/Tag.svelte';
+	import TokenName from '$lib/components/tokens/TokenName.svelte';
 
 	const dispatch = createEventDispatcher();
 
@@ -225,7 +227,7 @@
 	<div class="container md:max-h-[26rem] pr-2 pt-1 overflow-y-auto overscroll-contain">
 		{#each tokens as token (`${token.network.id.description}-${token.id.description}`)}
 			<Card>
-				{token.name}
+				<TokenName {token} />
 
 				<TokenLogo slot="icon" color="white" {token} />
